requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.13       # FastAPI v0.110 still uses Pydantic v1 by default
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pymongo==4.6.3
python-dotenv==1.0.1
alembic==1.13.1
celery==5.3.6
redis==5.0.4
